refactor: Move the AniList client (as-is for now) to a self-contained provider package
