+++ src/core/__init__.py
@@ -1,6 +1,5 @@
 """Core Module Initialization."""
 
-from src.core.anilist import AniListClient
 from src.core.animap import AniMapClient
 from src.core.plex import PlexClient
 
@@ -8,7 +7,6 @@ from src.core.bridge import BridgeClient  # isort:skip
 from src.core.sched import SchedulerClient
 
 __all__ = [
-    "AniListClient",
     "AniMapClient",
     "BridgeClient",
     "PlexClient",
