+++ src/providers/anilist/__init__.py
@@ -1,0 +1,1 @@
+"""AniList provider package."""
