+++ src/providers/anilist/list.py
@@ -1,0 +1,1 @@
+"""AniList list provider implementation."""
