+++ src/providers/plex/library.py
@@ -1,0 +1,1 @@
+"""Plex library provider implementation."""
