+++ tests/core/test_anilist.py
@@ -5,7 +5,6 @@ from pathlib import Path
 
 import pytest
 
-from src.core.anilist import AniListClient
 from src.exceptions import AniListFilterError
 from src.models.schemas.anilist import (
     Media,
@@ -14,6 +13,7 @@ from src.models.schemas.anilist import (
     User,
     UserOptions,
 )
+from src.providers.anilist.client import AniListClient
 
 
 @pytest.mark.asyncio
